feat: add dynamic template generation
