+++ .eslintrc.js
@@ -7,4 +7,7 @@ module.exports = {
       rootDir: ["apps/*/"],
     },
   },
+  env: {
+    node: true,
+  },
 };
