+++ package.json
@@ -9,14 +9,15 @@
     "changeset": "changeset",
     "version-packages": "changeset version",
     "release": "turbo run build --filter=docs^... && changeset publish",
-    "package:rollup": "node tasks/generate-rollup-template.js"
+    "package": "node tasks/generate-template.js"
   },
   "devDependencies": {
     "@changesets/cli": "^2.25.2",
     "eslint": "^8.29.0",
     "eslint-config-mykit": "0.0.0",
     "prettier": "^2.8.0",
-    "turbo": "latest"
+    "turbo": "latest",
+    "prompts": "^2.4.2"
   },
   "packageManager": "yarn@1.22.11",
   "workspaces": [
