+++ templates/custom-template/template.json
@@ -1,0 +1,21 @@
+{
+  "version": "0.0.0",
+  "main": "dist/cjs/index.js",
+  "module": "dist/esm/index.js",
+  "types": "dist/index.d.ts",
+  "license": "MIT",
+  "scripts": {
+    "build": "rollup -c",
+    "lint": "eslint \"src/**/*.ts*\"",
+    "clean": "rm -rf .turbo && rm -rf node_modules && rm -rf dist"
+  },
+  "devDependencies": {
+    "@types/react": "^18.0.9",
+    "@types/react-dom": "^18.0.4",
+    "eslint": "^8.15.0",
+    "eslint-config-mykit": "*",
+    "react": "^18.1.0",
+    "typescript": "^4.5.3",
+    "@rollup/plugin-typescript": "^11.1.2"
+  }
+}
