+++ templates/custom-template/template/.eslintrc.js
