+++ templates/custom-template/template/tsconfig.json
