+++ templates/rollup-typescript-template/index.js
