+++ templates/rollup-typescript-template/package.json
