+++ templates/rollup-typescript-template/template.json
