+++ templates/rollup-typescript-template/template/rollup.config.mjs
