+++ templates/rollup-typescript-template/template/src/example/Example.tsx
