+++ templates/rollup-typescript-template/template/src/example/index.ts
