+++ templates/rollup-typescript-template/template/src/index.ts
